+++ src/itext.c
@@ -64,7 +64,7 @@ itext_info(iObject *iobject, VipsBuf *buf, int indent)
 {
 	iText *itext = ITEXT(iobject);
 
-	vips_buf_appendf(buf, "%*c", indent, ' ');
+	vips_buf_space(buf, indent);
 	vips_buf_appends(buf, _("Formula"));
 	vips_buf_appendf(buf, ": %s\n", itext->formula);
 }