+++ src/main-snip.c
@@ -341,6 +341,9 @@ main(int argc, char **argv)
 	g_unsetenv("VIPSHOME");
 #endif /*FLATPAK*/
 
+	// disable DoS limits on libvips 8.19+
+    g_setenv("VIPS_UNLIMITED", "1", TRUE);
+
 	if (VIPS_INIT(argv[0]))
 		vips_error_exit("unable to start libvips");
 
@@ -359,20 +362,13 @@ main(int argc, char **argv)
     g_option_group_set_translation_domain(main_group, GETTEXT_PACKAGE);
     g_option_context_set_main_group(context, main_group);
 
-    GError *error = NULL;
+	/* Don't look for errors, we pass arg processing on to .defs.
+	 */
 #ifdef G_OS_WIN32
-    if (!g_option_context_parse_strv(context, &argv, &error))
+    (void) g_option_context_parse_strv(context, &argv, NULL);
 #else  /*!G_OS_WIN32*/
-    if (!g_option_context_parse(context, &argc, &argv, &error))
+    (void) g_option_context_parse(context, &argc, &argv, NULL);
 #endif /*G_OS_WIN32*/
-    {
-        if (error) {
-            fprintf(stderr, "%s\n", error->message);
-            g_error_free(error);
-        }
-
-        vips_error_exit("try \"%s --help\"", g_get_prgname());
-    }
 
     g_option_context_free(context);
 
@@ -483,7 +479,7 @@ main(int argc, char **argv)
 			main_error_exit("%s", _("no \"main\" found"));
 	}
 
-    //main_shutdown();
+    main_shutdown();
 
     return 0;
 }