+++ src/managed.c
@@ -159,16 +159,16 @@ managed_info(iObject *iobject, VipsBuf *buf, int indent)
 #ifdef DEBUG
 	Managed *managed = MANAGED(iobject);
 
-	vips_buf_appendf(buf, "%*cmanaged-object %p\n", indent, ' ',
-		managed);
-	vips_buf_appendf(buf, "%*cmanaged->count = %d\n", indent, ' ',
-		managed->count);
-	vips_buf_appendf(buf, "%*cmanaged->marked = %d\n", indent, ' ',
-		managed->marked);
+	vips_buf_space(buf, indent);
+	vips_buf_appendf(buf, "managed-object %p\n", managed);
+	vips_buf_space(buf, indent);
+	vips_buf_appendf(buf, "managed->count = %d\n", managed->count);
+	vips_buf_space(buf, indent);
+	vips_buf_appendf(buf, "managed->marked = %d\n", managed->marked);
 #endif /*DEBUG*/
 
-	vips_buf_appendf(buf, "%*c%s %p", indent, ' ',
-		G_OBJECT_TYPE_NAME(iobject), iobject);
+	vips_buf_space(buf, indent);
+	vips_buf_appendf(buf, "%s %p", G_OBJECT_TYPE_NAME(iobject), iobject);
 }
 
 static void