+++ src/managedfile.c
@@ -55,12 +55,13 @@ managedfile_info(iObject *iobject, VipsBuf *buf, int indent)
 {
 	Managedfile *managedfile = MANAGEDFILE(iobject);
 
-	vips_buf_appendf(buf, "%*cmanagedfile->fp = %p\n", indent, ' ',
-		managedfile->file->fp);
-	vips_buf_appendf(buf, "%*cmanagedfile->file->filename = %s\n", indent, ' ',
+	vips_buf_space(buf, indent);
+	vips_buf_appendf(buf, "managedfile->fp = %p\n", managedfile->file->fp);
+	vips_buf_space(buf, indent);
+	vips_buf_appendf(buf, "managedfile->file->filename = %s\n",
 		managedfile->file->fname);
-	vips_buf_appendf(buf, "%*cmanagedfile->file->last_errno = %d\n",
-		indent, ' ',
+	vips_buf_space(buf, indent);
+	vips_buf_appendf(buf, "managedfile->file->last_errno = %d\n",
 		managedfile->file->last_errno);
 
 	IOBJECT_CLASS(managedfile_parent_class)->info(iobject, buf, indent);