+++ src/managedgvalue.c
@@ -56,9 +56,9 @@ managedgvalue_info(iObject *iobject, VipsBuf *buf, int indent)
 	Managedgvalue *managedgvalue = MANAGEDGVALUE(iobject);
 	char *value_str;
 
+	vips_buf_space(buf, indent);
 	value_str = g_strdup_value_contents(&managedgvalue->value);
-	vips_buf_appendf(buf, "%*cmanagedgvalue->value = %s\n", indent, ' ',
-		value_str);
+	vips_buf_appendf(buf, "managedgvalue->value = %s\n", value_str);
 	g_free(value_str);
 
 	IOBJECT_CLASS(managedgvalue_parent_class)->info(iobject, buf, indent);